requirements.txt
# PDF to JSON Processing Tool - Dependencies

# Google Gemini API
google-generativeai>=0.3.0

# Environment variables management
python-dotenv>=1.0.0
